refactor(feature/donkey/#99): useMutation 옵션 타입 분리
