+++ yeoreum/src/types/mutation.ts
@@ -1,0 +1,15 @@
+import { AxiosResponse } from 'axios';
+
+type OnSuccess =
+  | ((
+      data: AxiosResponse<any, any>,
+      variables: void,
+      context: unknown,
+    ) => unknown)
+  | undefined;
+
+type OnError =
+  | ((error: any, variables: void, context: unknown) => unknown)
+  | undefined;
+
+export type { OnSuccess, OnError };
